Split Chip and RemovableChip
